# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdTech Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects/{subjectId}/chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters in a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter (admin only)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterInput"
      responses:
        "201":
          description: Created chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{chapterId}:
    get:
      operationId: getChapter
      tags: [chapters]
      summary: Get chapter with topics
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chapter detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterDetail"
    patch:
      operationId: updateChapter
      tags: [chapters]
      summary: Update chapter (admin only)
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterUpdate"
      responses:
        "200":
          description: Updated chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"

  # Topics
...
  /progress/subject/{subjectId}:
    get:
      operationId: getSubjectProgress
      tags: [progress]
      summary: Get progress for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subject progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectProgress"

  # Exams
...
components:
  schemas:
...
    Chapter:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        totalTopics:
          type: integer
        completedTopics:
          type: integer
        progressPercent:
          type: number
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        chapterTestUnlocked:
          type: boolean
        notesUploadUnlocked:
          type: boolean
      required:
        - id
        - subjectId
        - name
        - order
        - totalTopics
        - completedTopics
        - progressPercent
        - gateStatus
        - chapterTestUnlocked
        - notesUploadUnlocked
    ChapterInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
      required:
        - name
    ChapterUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        order:
          type: integer
    ChapterDetail:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        progressPercent:
          type: number
        chapterTestUnlocked:
          type: boolean
        notesUploadUnlocked:
          type: boolean
        topics:
          type: array
          items:
            $ref: "#/components/schemas/Topic"
      required:
        - id
        - subjectId
        - name
        - order
        - gateStatus
        - progressPercent
        - chapterTestUnlocked
        - notesUploadUnlocked
        - topics

    # Topics
    Topic:
      type: object
      properties:
        id:
          type: string
        chapterId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        telegramChatId:
          type: ["string", "null"]
        telegramMessageId:
          type: ["string", "null"]
        lectureQuizPassed:
          type: boolean
        dppCompleted:
          type: boolean
        pyqCompleted:
          type: boolean
        topicTestPassed:
          type: boolean
        isComplete:
          type: boolean
        gateStatus:
          type: string
          enum: [locked, unlocked, completed]
        lectureClickCount:
          type: integer
      required:
        - id
        - chapterId
        - name
        - order
        - lectureQuizPassed
        - dppCompleted
        - pyqCompleted
        - topicTestPassed
        - isComplete
        - gateStatus
        - lectureClickCount
...
    SubjectProgress:
      type: object
      properties:
        subjectId:
          type: string
        subjectName:
          type: string
        progressPercent:
          type: number
        completedTopics:
          type: integer
        totalTopics:
          type: integer
        weakTopics:
          type: array
          items:
            type: string
      required:
        - subjectId
        - subjectName
        - progressPercent
        - completedTopics
        - totalTopics
        - weakTopics

    # Exams
...